fix(keycloak): remove code duplication
